Reliable writing records method to DB
First, try to avoid using loops and add all together at once.
This is mainly for speed performance.

Then, try splitting the input records into data chunks, and
add the chunks all at once (to avoid loops once again).

If there happens to be some error when trying to commit
a chunk of records (perhaps due to duplicates), final fallback
is to write the records of that one bulk one-by-one.

This process is much slower but ensures that the records do
get written onto the DB without any fail.